Giving up. Read commit message:
There is wayyyyyy to many things wrong with the setup that I see no end
to fixing it. the benchmark depends on an old version of Dicourse
which doesn't work with 2.7 and there is a lot of fallout from that

I got it to the point where start.rb would boot until it gets a 403
from the server. and that's on 2.6

For 2.7, it fails because of a weird bundler issue. Bundler 1.17.3
is installed, but it complains that it's missing.